+++ magenta/interfaces/midi/BUILD
@@ -70,6 +70,6 @@ py_binary(
         ":midi_interaction",
         "//magenta",
         "//magenta/models/melody_rnn:melody_rnn_sequence_generator",
-        "//magenta/models/polyphonic_rnn:polyphony_sequence_generator",
+        "//magenta/models/polyphony_rnn:polyphony_sequence_generator",
     ],
 )
